backend/requirements.txt
# Project Cerberus Dependencies
# Simple, minimal dependencies for a 3rd-year student project

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.9.2
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.1
